fix: pytest duplicated addoptions on Windows Pyinstaller one-file build
